agents/community/langgraph-agentic-rag/pyproject.toml
[tool.poetry]
name = "langgraph_agentic_rag"
version = "0.2.0"
description = "A LangGraph Agentic RAG app using OpenAI-compatible APIs. Supports retrieval-augmented generation with customizable vector stores."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "langgraph_agentic_rag", from = "src" }
]


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain-openai = ">=1.0.0"
openai = ">=1.0.0"
langgraph = ">=1.0.0"
langgraph-prebuilt = ">=1.0.0"
langchain-core = ">=1.0.0"
langchain = ">=1.0.0"
langchain-community = ">=0.3.0"
faiss-cpu = "^1.7.4"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.0"
pydantic = ">=2.0.0"
typing-extensions = ">=4.5.0"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"